Introduce support for APL-A
APL-A is Alexa Presentation Language for audio effects. It is similar to
APL in its structure (directives, data sources, documents etc.). However,
it is different from APL in that there is no concept of supported interfaces.
It is supported on all Alexa devices.

In this commit, we add support for validating APLA directives and also added
a test.

I was divided on whether we should just add APL-A to existing APL validator
or if we should create a compeltely independent validator for APL-A. For now
I added it to the existing APLValidator but open to feedback.

https://github.com/taimos/ask-sdk-test/issues/20