Support withShouldEndSession having an undefined value.
shouldEndSession has a special meaning when it is undefined. It is
usually used on APL devices to leave the visuals showing without opening
up the mic.

I modified most tests to explicity end the session.

The touch handler test was written to leave the shouldEndSession in an undefined
state to test the changes in this commit.

https://developer.amazon.com/en-US/docs/alexa/custom-skills/manage-skill-session-and-session-attributes.html#screen-session