fix: Simulate a device that supports video (#37)
Using the condition `.withInterfaces({ video: true }).build()` to build a request does not make the `handlerInput.requestEnvelope.context.Viewport.video` to be _truthy_ in order to simulate a device that does support video. It only changes the value of `handlerInput.requestEnvelope.context.System.device.supportedInterfaces.VideoApp`.

The `Viewport.video` property represents an object including an array of `codecs` that contains the available technologies for playing video on the device. If it does not support playing videos, the `Viewport.video` property is not present. To simulate a device that does support playing video it should be enough to make that property exist (or be _truthy_). In this PR, this property has been set to contain an array with all the possible `codecs` that the output device could support.

Here's the related issue:
- https://github.com/taimos/ask-sdk-test/issues/31